chore: Organize holdings service better
