+++ src/routes/investments.route.ts
@@ -9,8 +9,8 @@ import {
 
 import {
   loadHoldingsList,
-  addHolding,
-} from '@services/investments/holdings.service';
+  createHolding,
+} from '@services/investments/holdings';
 import {
   createInvestmentTransaction,
   getInvestmentTransactions,
@@ -34,7 +34,7 @@ router.post('/holdings', authenticateJwt, async (req, res) => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const { id: userId } = req.user as any;
   const { accountId, securityId } = req.body;
-  const data = await addHolding({ userId, accountId, securityId });
+  const data = await createHolding({ userId, accountId, securityId });
 
   return res.status(200).json({
     status: API_RESPONSE_STATUS.success,
