+++ src/services/investments/holdings/create-holding.service.ts
@@ -2,51 +2,10 @@ import { GenericSequelizeModelAttributes } from '@common/types';
 import { connection } from '@models/index';
 
 import Account from '@models/Accounts.model';
-import User from '@models/Users.model';
 import Holdings from '@models/investments/Holdings.model';
 import Security from '@models/investments/Security.model';
 
-export async function loadHoldingsList(
-  { userId }: { userId: number },
-  { transaction }: GenericSequelizeModelAttributes = {},
-) {
-  const isTxPassedFromAbove = transaction !== undefined;
-  transaction = transaction ?? (await connection.sequelize.transaction());
-
-  try {
-    const holdings = await Holdings.findAll({
-      include: [
-        {
-          model: Account,
-          required: true,
-          where: { userId }, // Direct filtering by userId on the Account
-          include: [
-            {
-              model: User,
-              required: true,
-            },
-          ],
-        },
-      ],
-      transaction,
-    });
-
-    if (!isTxPassedFromAbove) {
-      await transaction.commit();
-    }
-
-    return holdings;
-  } catch (err) {
-    if (!isTxPassedFromAbove) {
-      await transaction.rollback();
-    }
-
-    throw err;
-  }
-}
-
-// TODO: e2e tests
-export async function addHolding(
+export async function createHolding(
   {
     userId,
     accountId,
@@ -106,7 +65,3 @@ export async function addHolding(
     throw err;
   }
 }
-
-// TODO: Update existing via InvestmentTransactions
-// TODO: Delete existing
-// TODO: Check that existing is being deleted if Account is also deleted
