+++ src/services/investments/holdings/index.ts
@@ -1,0 +1,6 @@
+export * from './load-holdings.service';
+export * from './create-holding.service';
+
+// TODO: Update existing via InvestmentTransactions
+// TODO: Delete existing
+// TODO: Check that existing is being deleted if Account is also deleted
